Refactor: Response class 패키지 경로 변경 (#21)
