+++ src/main/java/org/sesac/slopedbe/auth/model/vo/JwtResponse.java
@@ -1,4 +1,4 @@
-package org.sesac.slopedbe.auth.model.dto;
+package org.sesac.slopedbe.auth.model.vo;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
